snapshot command working
