+++ cli/src/lib/rpc-client.ts
@@ -5,11 +5,12 @@ import {
   decodeFunctionResult,
   encodeFunctionData,
   type Hex,
-  http,
+  http, numberToHex,
 } from "viem";
 import type { TypeOf, ZodType } from "zod";
 import type { PublicClient, HttpTransport } from "viem";
 import { memoryDiffParser, type MemoryDiffRaw } from "../schema/rpc";
+import {getStorageAt} from "viem/actions";
 
 const L1_DEFAULT_URLS = {
   mainnet: "https://ethereum-rpc.publicnode.com",
@@ -59,6 +60,19 @@ export class RpcClient {
 
   async getByteCode(addr: Hex): Promise<Hex | undefined> {
     return this.viemClient.getBytecode({ address: addr });
+  }
+
+  async storageRead(addr: Hex, position: bigint): Promise<Hex> {
+    const readedValue = await getStorageAt(this.viemClient, {
+      address: addr,
+      slot: numberToHex(position, { size: 32 })
+    })
+
+    if (!readedValue) {
+      throw new Error("Error reading storage")
+    }
+  
+    return readedValue
   }
 
   async contractRead<T extends ZodType>(
