+++ cli/src/lib/storage/rpc-storage-snapshot.ts
@@ -1,8 +1,19 @@
 import type {StorageSnapshot} from "./storage-snapshot";
-import type {Option} from "nochoices";
+import {Option} from "nochoices";
+import type {RpcClient} from "../rpc-client";
+import {type Hex, hexToBytes} from "viem";
 
 export class RpcStorageSnapshot implements StorageSnapshot {
+  private rpc: RpcClient;
+  private addr: Hex
+
+  constructor(rpc: RpcClient, addr: Hex) {
+    this.rpc = rpc
+    this.addr = addr
+  }
+
   async at(pos: bigint): Promise<Option<Buffer>> {
-    return Promise.reject(new Error("Attempting to retrieve an attempt"));
+    const hex = await this.rpc.storageRead(this.addr, pos);
+    return Option.Some(Buffer.from(hexToBytes(hex)))
   }
 }
